requirements.txt
pandas
torch
torchvision
numpy
RPi.GPIO